requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.12
jinja2==3.1.4
requests==2.32.3
python-dotenv==1.0.1
numpy==2.1.3
opencv-python==4.10.0.84
mediapipe==0.10.14
